+++ src/Sulu/Bundle/MarkupBundle/Tag/TagInterface.php
@@ -20,8 +20,19 @@ interface TagInterface
      * Returns new tag with given attributes.
      *
      * @param array $attributesByTag attributes array of each tag occurrence.
+     * @param string $locale
      *
      * @return array Tag array to replace all occurrences.
      */
-    public function parseAll($attributesByTag);
+    public function parseAll(array $attributesByTag, $locale);
+
+    /**
+     * Returns array of validity.
+     *
+     * @param array $attributesByTag
+     * @param string $locale
+     *
+     * @return array Tag array which indicates validity of each tag.
+     */
+    public function validateAll(array $attributesByTag, $locale);
 }